Rename test-utils folder to utils.
