+++ tests/frontend/components/MeshModal.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from 'vitest'
 import MeshModal from '@/components/MeshModal.vue'
-import { mountModalWithStubs } from '../test-utils/mountWithGlobalStubs'
+import { mountModalWithStubs } from '../utils/mountWithGlobalStubs'
 
 function mountMeshModal(customProps = {}) {
   return mountModalWithStubs(MeshModal, {
