+++ tests/frontend/utils/mountWithGlobalStubs.ts
